+++ poem-openapi/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.3.17] 2022-3-23
+
+- Implement `From<T: ApiResponse>` for `poem::Error` so that it can be used as error type.
+
 # [1.3.14] 2022-3-10
 
 - Add support for use multiple methods on a single endpoint. [#229](https://github.com/poem-web/poem/discussions/229)